Fix negative progress percentage in matrix generation
- Use long long to prevent integer overflow in progress calculation
- Fixes issue with large matrices (>11000x11000) showing negative percentages

+++ src/io/io.cu
@@ -334,8 +334,8 @@ int write_matrix_market_stencil5 (int n, const char* filename){
 	fprintf(f, "%d %d %d\n", grid_size, grid_size, nnz);
 
 	// Write values with progress indication
-	int total_points = n * n;
-	int progress_step = total_points / 100; // Print every 1%
+	long long total_points = (long long)n * n;
+	long long progress_step = total_points / 100; // Print every 1%
 	if (progress_step == 0) progress_step = 1; // Ensure progress for small matrices
 	
 	printf("Writing matrix entries: 0%%");
@@ -344,11 +344,11 @@ int write_matrix_market_stencil5 (int n, const char* filename){
 	for (int row = 0; row < n; row++) {
 		for (int col = 0; col < n; col++) {
 			int idx = row * n + col + 1;  // 1-based index
-			int current_point = row * n + col;
+			long long current_point = (long long)row * n + col;
 			
 			// Progress indicator - update every 1% or every 1000 points (whichever is smaller)
 			if (current_point % progress_step == 0 || current_point % 1000 == 0) {
-				int percent = (current_point * 100) / total_points;
+				int percent = (int)((current_point * 100) / total_points);
 				printf("\rWriting matrix entries: %d%%", percent);
 				fflush(stdout);
 			}
